Refactoring has been done.

+++ WebApi/Application/DirectorOperations/Commands/DeleteDirector/DeleteDirectorCommand.cs
@@ -1,6 +1,7 @@
 ﻿using System.Linq;
 using System;
 using WebApi.DbOperations;
+using Microsoft.EntityFrameworkCore;
 
 namespace WebApi.Application.DirectorOperations.Commands.DeleteDirector
 {
@@ -19,6 +20,11 @@ namespace WebApi.Application.DirectorOperations.Commands.DeleteDirector
             var director = _dbContext.Directors.SingleOrDefault(x => x.Id == DirectorId);
             if (director is null)
                 throw new InvalidOperationException("Silnecek Director bulunamadı.");
+
+            bool isDirectingAnyMovie = _dbContext.Movies.Include(movie => movie.Director).Any(movie => movie.IsActive && movie.Director.Id == director.Id);
+
+            if (isDirectingAnyMovie)
+                throw new InvalidOperationException("Bu yönetmen bir film yönetmektedir, şu anda silinemez.");
 
             _dbContext.Directors.Remove(director);
             _dbContext.SaveChanges();
